+++ modules/pota.cc
@@ -174,7 +174,7 @@ void pota_spots(ScreenFrame& panel, TTF_Font* font) {
        spots_raw.clear();
     } else {
         spots_raw.clear();
-        std::string sanitized(json_spots, data_size);
+        std::string sanitized(json_spots, static_cast<size_t>(data_size));
         spots_raw.str(sanitized);
         if(json_spots) {
              free (json_spots);
